+++ packages/terra-clinical-item-display/tests/jest/ItemDisplay.test.jsx
@@ -11,7 +11,7 @@ it('should render a default component', () => {
 
 it('should render with text', () => {
   const itemDisplay = shallowWithIntl(<Display text="Display" />).shallow();
-  expect(itemDisplay.find('.text').prop('children')).toEqual('Display');
+  expect(itemDisplay.find('.text').prop('children')).toEqual(<span>Display</span>);
   expect(itemDisplay).toMatchSnapshot();
 });
 
@@ -66,45 +66,52 @@ it('should render a disabled display', () => {
 
 it('should render a display with the primary textStyle on text', () => {
   const itemDisplay = shallowWithIntl(<Display text="Display" textStyle="primary" />).shallow();
+  expect(itemDisplay.contains(<span>Display</span>)).toBe(true);
   expect(itemDisplay.find('.text').prop('className')).toEqual('text');
   expect(itemDisplay).toMatchSnapshot();
 });
 
 it('should render a display with the attention textStyle on text', () => {
   const itemDisplay = shallowWithIntl(<Display text="Display" textStyle="attention" />).shallow();
+  expect(itemDisplay.contains(<span>Display</span>)).toBe(true);
   expect(itemDisplay.find('.text').prop('className')).toContain('attention');
   expect(itemDisplay).toMatchSnapshot();
 });
 
 it('should render a display with the secondary textStyle on text', () => {
   const itemDisplay = shallowWithIntl(<Display text="Display" textStyle="secondary" />).shallow();
+  expect(itemDisplay.contains(<span>Display</span>)).toBe(true);
   expect(itemDisplay.find('.text').prop('className')).toContain('secondary');
   expect(itemDisplay).toMatchSnapshot();
 });
 
 it('should render a display with the strikeThrough textStyle on text and default textStyleMeaning provided', () => {
   const itemDisplay = shallowWithIntl(<Display text="Display" textStyle="strikeThrough" />).shallow();
-  expect(itemDisplay.find('span').prop('aria-label')).toEqual('Terra.item-display.textStyleMeaningStrikethrough, Display, Terra.item-display.textStyleMeaningStrikethroughEnd');
+  expect(itemDisplay.contains(<span>Display</span>)).toBe(true);
+  expect(itemDisplay.find('span').first().prop('aria-label')).toEqual('Terra.item-display.textStyleMeaningStrikethrough, Display, Terra.item-display.textStyleMeaningStrikethroughEnd');
   expect(itemDisplay.find('.text').prop('className')).toContain('strike-through');
   expect(itemDisplay).toMatchSnapshot();
 });
 
 it('should render a display with the strikeThrough textStyle on text and aria label based on provided textStyleMeaning prop', () => {
   const itemDisplay = shallowWithIntl(<Display text="Display" textStyle="strikeThrough" textStyleMeaning="Update" />).shallow();
-  expect(itemDisplay.find('span').prop('aria-label')).toEqual('Update, Display, Terra.item-display.textStyleMeaningEnd');
+  expect(itemDisplay.contains(<span>Display</span>)).toBe(true);
+  expect(itemDisplay.find('span').first().prop('aria-label')).toEqual('Update, Display, Terra.item-display.textStyleMeaningEnd');
   expect(itemDisplay.find('.text').prop('className')).toContain('strike-through');
   expect(itemDisplay).toMatchSnapshot();
 });
 
 it('should render a display with a non-strikeThrough textStyle on text and aria label based on provided textStyleMeaning prop', () => {
   const itemDisplay = shallowWithIntl(<Display text="Display" textStyle="attention" textStyleMeaning="Important" />).shallow();
-  expect(itemDisplay.find('span').prop('aria-label')).toEqual('Important, Display, Terra.item-display.textStyleMeaningEnd');
+  expect(itemDisplay.contains(<span>Display</span>)).toBe(true);
+  expect(itemDisplay.find('span').first().prop('aria-label')).toEqual('Important, Display, Terra.item-display.textStyleMeaningEnd');
   expect(itemDisplay.find('.text').prop('className')).toContain('attention');
   expect(itemDisplay).toMatchSnapshot();
 });
 
 it('should render a display with the strong textStyle on text', () => {
   const itemDisplay = shallowWithIntl(<Display text="Display" textStyle="strong" />).shallow();
+  expect(itemDisplay.contains(<strong>Display</strong>)).toBe(true);
   expect(itemDisplay.find('.text').prop('className')).toContain('strong');
   expect(itemDisplay).toMatchSnapshot();
 });